fix(admin): satisfy web lint rules for email composer page
Split the composer hook from the editor component and inline audience loading to match existing admin page patterns.
